use stack trace when reporting error feature line
- move TryFindFeatureSource to extensions
