+++ core/src/Baked/Domain/DomainExtensions.cs
@@ -7,6 +7,7 @@ using Baked.Testing;
 using Microsoft.Extensions.DependencyInjection;
 using Shouldly;
 using System.ComponentModel;
+using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
 using System.Runtime.CompilerServices;
@@ -61,6 +62,45 @@ public static class DomainExtensions
         public static DiagnosticCode OrderOutOfBounds => new(303, "order-out-of-bounds");
         public static DiagnosticCode UndefinedLevel => new(304, "undefined-level");
         public static DiagnosticCode InvalidOrder => new(305, "invalid-order");
+    }
+
+    extension(StackTrace stackTrace)
+    {
+        public bool TryFindFeatureSource([NotNullWhen(true)] out string? source)
+        {
+            source = null;
+
+            var frames = stackTrace.GetFrames();
+            var featureFrame =
+                frames.FirstOrDefault(f => f.GetMethod()?.ReflectedType?.DeclaringType?.Name.EndsWith("Feature") == true) ??
+                frames.FirstOrDefault(f => f.GetMethod()?.DeclaringType?.Name.EndsWith("Feature") == true);
+            if (featureFrame is null) { return false; }
+
+            source =
+                featureFrame.GetMethod()?.ReflectedType?.DeclaringType?.Name ??
+                featureFrame.GetMethod()?.DeclaringType?.Name ??
+                string.Empty;
+
+            var fileName = featureFrame.GetFileName();
+            if (fileName is null)
+            {
+                source += " [bold red]⠕[/]";
+
+                return true;
+            }
+
+            var title = source;
+            var url = new Uri(fileName).AbsoluteUri;
+            source = $"[link={url}]{title}[/]";
+
+            var lineNumber = featureFrame.GetFileLineNumber();
+            if (lineNumber > 0)
+            {
+                source += $":{lineNumber}";
+            }
+
+            return true;
+        }
     }
 
     extension(ApplicationContext application)
